Create character development
he Graphic user interface has been made in such a way that the User Experience feels as natural and easy as possible. Some new innovative features are for example the green life bar on the top left corner of the window; the map that can be actually seen when found during the game; an accurate management of weapons and tools that can be found within the game and many other features that you will discover playing this adventure.